+++ libgdf/src/Reader.cpp
@@ -100,7 +100,7 @@ namespace gdf
     {
         resetCache( );
         m_record_cache.clear( );
-        size_t num_records = m_header.getMainHeader_readonly().get_num_datarecords();
+		size_t num_records = boost::numeric_cast<size_t>( m_header.getMainHeader_readonly().get_num_datarecords() );
         m_record_cache.resize( num_records, NULL );
     }
